Update Login Page
- Moved LoginPage.js/.css into Pages folder
- Installed router and mapped in index.js
- added link in LoginPage.js (now it's functional)
