+++ chatapplication/src/App.js
@@ -1,6 +1,5 @@
 import './App.css';
 import Registration from './Pages/Registration';
-import './App.css';
 import PopUp from "./components/PopUp";
 function App() {
   return (
