+++ chatapplication/src/Pages/LoginPage.css
@@ -1,5 +1,3 @@
-@import "./components/palette.css";
-
 .vertical-line {
     position: absolute;
     width: 400px;
@@ -8,6 +6,11 @@
 
     border: 0.5px solid #282727;
     transform: rotate(90deg);
+}
+
+.body{
+    color: #000000;
+    
 }
 
 .card{
@@ -25,7 +28,15 @@
     border: 1px solid #737971;
 }
 
-.card-column{
+.title{
+    font-family: 'Roboto';
+    font-style: normal;
+    font-weight: 700;
+    font-size: 30px;
+    line-height: 22px;
+}
+
+.card--column{
     display:flex;
     flex-direction: column;
     justify-content: center;
